refactor(primitives): extract authenticator key set module
